WEEK 7 6.0 with invoice and invoice_products repository functionality

+++ BACKEND/WEEK_7_BACKEND/routes/contact_routes.py
@@ -1,36 +1,28 @@
 
 from flask import request, jsonify, Blueprint
 from repositories.contact_repository import ContactRepository
-from services.jwt_manager import JwtManager
-from services.decorators import roles_required
+from services.decorators import roles_required,get_jwt_identity
 from pathlib import Path
-
-base_path = Path(__file__).resolve().parent.parent
-with open(base_path / "keys" / "private.pem", "rb") as f:
-    private_key = f.read()
-
-with open(base_path / "keys" / "public.pem", "rb") as f:
-    public_key = f.read()
-
-jwt_manager=JwtManager(private_key=private_key,public_key=public_key)
 
 contact_repo = ContactRepository()
 contacts_bp=Blueprint("contacts",__name__)
 
 
-
 @contacts_bp.route("/", methods=["POST"])
-@roles_required("administrator")
+@roles_required("administrator","user")
 def create_contact():
     try:
+        user_data=get_jwt_identity()
         contact_data=request.get_json()
+
         required_fields = ["user_id", "name","phone","email"]
         missing_fields = [field for field in required_fields if field not in contact_data]
-
         if missing_fields:
             return jsonify({"error": f"Missing fields: {', '.join(missing_fields)}"}), 400
         
-        
+        if user_data["role"]=="user" and user_data["sub"]!=str(contact_data["user_id"] ):
+            return jsonify({"error": "Access denied"}), 403
+            
         new_contact=contact_repo.create(**contact_data)
 
         if new_contact:
@@ -41,11 +33,14 @@ def create_contact():
 
 
 @contacts_bp.route("/", methods=["GET"])
-@roles_required("administrator")
+@roles_required("administrator","user")
 def get_all_contacts():
     try:
-        
-        data_contacts=contact_repo.get_all()
+        user_data=get_jwt_identity()
+        if user_data["role"]=="user":
+            data_contacts=contact_repo.get_by_user(user_data["sub"])
+        else:
+            data_contacts=contact_repo.get_all()
         
         if not data_contacts:
             return jsonify({"data": [], "message": "No contacts found"}), 404
@@ -72,14 +67,17 @@ def get_by_id(identifier):
     
 
 @contacts_bp.route("/<identifier>", methods=["DELETE"])
-@roles_required("administrator")
-def delete_product(identifier):
+@roles_required("administrator","user")
+def delete_contact(identifier):
     try:
-        
-        data_contact=contact_repo.delete(identifier)
-
-        if data_contact is None:
+        user_data=get_jwt_identity()
+        user_id=contact_repo.get_user_id(identifier)
+        if user_id is None:
             return jsonify({"data": [], "message": "No contact found"}), 404
+        if user_data["role"]=="user" and user_data["sub"]!=str(user_id) :
+            return jsonify({"error": "Access denied"}), 403
+
+        deleted_contact=contact_repo.delete(identifier)
 
         return jsonify({
             "message": f"Contact with ID {identifier} was deleted successfully"
@@ -91,22 +89,31 @@ def delete_product(identifier):
     
 
 @contacts_bp.route("/<identifier>", methods=["PUT"])
-@roles_required("administrator")
+@roles_required("administrator","user")
 def update_contact(identifier):
     try:
+        user_data = get_jwt_identity()
+        data_contact = request.get_json()
+
+
+        contact_owner_id = contact_repo.get_user_id(identifier)
+        if contact_owner_id is None:
+            return jsonify({"error": "Contact not found"}), 404
+
         
-        data_contact=request.get_json()
-    
-        updated_contact=contact_repo.update(
+        if user_data["role"] == "user" and user_data["sub"] !=str( contact_owner_id):
+            return jsonify({"error": "Access denied"}), 403
+
+        updated_contact = contact_repo.update(
             contact_id=identifier,
             name=data_contact.get("name"),
             phone=data_contact.get("phone"),
             email=data_contact.get("email")
         )
-        if not updated_contact:
-            return jsonify({"error": "Contact not found"}), 404
+
         return jsonify({
-            "message": f"Contact with ID {identifier} was updated successfully"
+            "message": f"Contact with ID {identifier} was updated successfully",
+            "contact": updated_contact.to_dict()
         }), 200
 
     except Exception as ex:
@@ -114,29 +121,3 @@ def update_contact(identifier):
         return jsonify({"message": "Error updating contact."}), 500
 
 
-# @users_bp.route("/multiple-cars", methods=["GET"])
-# def get_multiple_cars():
-#     try:
-        
-#         data_users=UserRepository.get_users_with_multiple_cars()
-        
-#         if not data_users:
-#             return jsonify({"data": [], "message": "No users found"}), 404
-        
-#         return jsonify({"data": data_users}), 200
-#     except Exception as e:
-#         return jsonify({"error": "Unexpected error", "details": str(e)}), 500
-    
-
-# @users_bp.route("/<identifier>/relations", methods=["GET"])
-# def get_cars_addresses_from_user(identifier):
-#     try:
-        
-#         data_users=UserRepository.get_cars_addresses_from_user(identifier)
-        
-#         if not data_users:
-#             return jsonify({"data": [], "message": "No users found"}), 404
-        
-#         return jsonify({"data": data_users}), 200
-#     except Exception as e:
-#         return jsonify({"error": "Unexpected error", "details": str(e)}), 500
